requirements.txt
flask-admin==1.0.9
flask-sqlalchemy==2.0
flask-login==0.2.11
flask-wtf==0.11
wtforms==2.0.2
flask-migrate==1.3.0
flask-script==2.0.5
flask-principal==0.4.0

